adding support for photographers and editors
I'm not sure we'll end up wanting to keep both of these, but it
shouldn't take Ana & Jen to come to a decision about it once they see
them in action. So, they're here for now. Also, I took uploads\ out of
my gitignore so we'll be working with the same backup file