Search and highlighting update
using "exact_words" field to index words from most fields with lightest
stemming (only singular/plural). This field allows AND search across all
relevant field contents, with better faceting. Keeping the other fields
as part of the search allows the highlighting to function. If we used
q.op to set the AND operator, this would force all terms to appear in
all fields searched (matching nothing). The current workaround for this
is a regex modification of the user's search input to manually add "AND"
between words that aren't already AND/OR/NOT operators, or quoted
phrases.