CDM export fixes (typo, problem HTML)
Fixed typo in export.php. Removed divs from exportZip.php would be added
to the end of the zip file rather than displaying as html. Besides
which, the record export success message is unneccessary (if successful,
the user will have a valid zip file; if not, the most likely failures
will result in an invalid zip file, in which case this success message
would be generated in error).